Splitt settings and dialog into correct packages to clearify the architecture.
